git-svn-id: svn://svn.code.sf.net/p/passwordsafe/code/trunk/pwsafe/pwsafe@1674 1f79f812-37fb-46fe-a122-30589dd2bf55

+++ corelib/PWSclipboard.cpp
@@ -9,6 +9,7 @@
 /** \file
  * Implementation file for the PWSclipboard class
  */
+#include "../stdafx.h"
 
 #include <afxole.h>
 #include "PWSclipboard.h"
@@ -62,7 +63,7 @@ PWSclipboard::ClearData()
     HANDLE hData = odo.GetGlobalData(CLIPBOARD_TEXT_FORMAT);
     if (hData != NULL) {
       LPCTSTR pData = (LPCTSTR)::GlobalLock(hData);
-      DWORD dwlength =  ::GlobalSize(hData) - sizeof(TCHAR); // less trailing null
+      SIZE_T dwlength =  ::GlobalSize(hData) - sizeof(TCHAR); // less trailing null
       if (dwlength < 1)
         return !m_set;
